parse expression for some arguments, simply deparse for others
For "calculation," "constraint," "relevance," and "repeat_count," if the
argument is an expression, then that expression gets evaluated.

For other arguments, if the argument is an expression (`symbol` or
`language`), then it just gets converted to a string representation.

This means that the following are equivalent:

  Ask(foo, SelectOne(yes_no), "Foo?", required = "yes")
  Ask(foo, SelectOne(yes_no), "Foo?", required = yes)

But the following are not:

  Ask(foo, SelectOne(yes_no), "Foo?", constraint = ". == bar")
  Ask(foo, SelectOne(yes_no), "Foo?", constraint = . == bar)

The latter parses the constraint into the expression ". = ${bar}" but
the former uses ". == bar" verbatim.  The latter will thus check if
"bar" has been defined before, and it will generate a valid XPath
expression.  But the former will generate an invalid XPath expression.