let `When` accept any number of arguments
There is no reason for `When` to expect the block in a single argument.
Since it expects no other arguments, it can reconstruct the block from
`...`.  This is consistent with what `ChoiceList` does.

This change does not break usage of `When` that passes the block in a
single argument.  So now both of the following are accepted:

  When(
    foo == 1,
    c(
      Ask(bar, SelectOne(yes_no), "Bar?"),
      Ask(baz, SelectOne(yes_no), "Baz?")
    )
  )

  When(
    foo == 1,
    Ask(bar, SelectOne(yes_no), "Bar?"),
    Ask(baz, SelectOne(yes_no), "Baz?")
  )